+++ CHANGELOG.md
@@ -8,6 +8,34 @@
 
 ### Changed
 
+## [v0.1.0] - 2015-04-16
+
+### Fixed
+
+- Selfupdate updates the actual running binary intead of the fixed /us/local/bin/cbd
+- SMTP default port is 25, to fix number conversion exception
+
+### Added
+
+- Command `init` creates Profile
+- Install cbd to a directory which is available on $PATH
+- Docker based test for the one-liner install from README.md: `make install-test`
+
+### Removed
+
+- `update-snap` command removed, replaced by parametrized `update`
+
+### Changed
+
+- Cloudbreak/Persicope/Uluwatu/Sultans Dcoker images upgraded to 0.4.x
+- Use the built in 'checksum' function instead of the external 'shasum' to generate secrets
+- Command `update` by default updates from latest Github release, parameter can point to branch on CircleCI
+- DOCKER_XXX env varibles are inherited, so they not needed in Profile
+- `generate` and compose specific commands are only available when `Profile` exists
+- `generate` command genertes docker-compose.yml **and** uaa.yml
+- `PRIVATE_IP` env var defaults to bridge IP (only PUBLC_IP is required in Profile)
+- use **sulans-bin** docker image istead of sultans
+
 ## [v0.0.9] - 2015-04-14
 
 ### Fixed
@@ -142,6 +170,6 @@ Added
 
 - help command added
 - version command added
-- Added --version 
+- Added --version
 - CircleCI build
 - Linux/Darwin binary releases on github